+++ packages/frontend/src/stdlib/theories.ts
@@ -1,7 +1,13 @@
 import { TheoryLibrary } from "../theory";
+import { compositionPattern } from "./generic_analyses";
 
 /** Standard library of double theories supported by the frontend. */
 export const stdTheories = new TheoryLibrary();
+
+stdTheories.addGenericModelAnalysis({
+    construct: compositionPattern,
+    when: (theory) => theory.theory.canInstantiateModels(),
+});
 
 stdTheories.add(
     {