+++ CHANGELOG.md
@@ -9,6 +9,14 @@ announcement and a blog post. Minor versions are not announced but allow
 features and fixes to be released with greater frequency. Minor versions often
 include notable new features.
 
+## [Unreleased]
+
+### Added
+
+- New analysis: visualize composition pattern of instantiated models as an
+  undirected wiring diagram
+  ([#1117](https://github.com/ToposInstitute/CatColab/pull/1117/))
+
 ## [v0.5.0](https://github.com/ToposInstitute/CatColab/releases/tag/v0.5.0) (2026-03-09)
 
 ### Added